🐛 fix: preserve project list workspace scope
