+++ src/features/Workspace/__tests__/workspaceAwarePath.test.ts
@@ -22,6 +22,7 @@ describe('buildWorkspaceAwarePath', () => {
     );
     expect(buildWorkspaceAwarePath('/group/group-1', 'acme')).toBe('/acme/group/group-1');
     expect(buildWorkspaceAwarePath('/project/project-1', 'acme')).toBe('/acme/project/project-1');
+    expect(buildWorkspaceAwarePath('/projects', 'acme')).toBe('/acme/projects');
   });
 
   it('prefixes deep agent and evaluation paths used by cross-page navigation', () => {
