+++ src/spa/router/desktopRouter.sync.test.tsx
@@ -146,10 +146,13 @@ describe('desktop router shared definition', () => {
   );
 
   it.each(mainAreaVariants)('%s exposes the projects view-all route', (_, factory) => {
-    const matches = matchRoutes(createMainAreaRoutes(factory), '/projects');
+    const personalMatches = matchRoutes(createMainAreaRoutes(factory), '/projects');
+    const workspaceMatches = matchRoutes(createMainAreaRoutes(factory), '/acme/projects');
 
-    expect(matches?.at(-1)?.route.index).toBe(true);
-    expect(matches?.at(-1)?.route.handle).toMatchObject({ meta: expect.any(Object) });
+    expect(personalMatches?.at(-1)?.route.index).toBe(true);
+    expect(personalMatches?.at(-1)?.route.handle).toMatchObject({ meta: expect.any(Object) });
+    expect(workspaceMatches?.at(-1)?.route.index).toBe(true);
+    expect(workspaceMatches?.at(-1)?.route.handle).toMatchObject({ meta: expect.any(Object) });
   });
 
   it.each(mainAreaVariants)(
